forgot to add file

+++ src/blant.h
@@ -75,7 +75,7 @@ extern unsigned long _known_orbit_count[2][13];
   #if int_width >= 28 && short_width >= 16
     typedef unsigned Gint_type; // at k=8, max lookup index is 2^28, so we need 32 bits...
     #define GINT_FMT "%u"
-    typedef unsigned Gordinal_type;
+    typedef unsigned short Gordinal_type;
     //typedef unsigned short Gordinal_type; //... and max numCanonicals is 12348 < 2^16, so 16 bits is sufficient (commented out b/c directed graphs require 18 bits)
     #define GORDINAL_FMT "%u"
     #define MAX_BINTREE_K 8
